Renamed command upload-assets to upload
